Solution to Calculate Employee Bonus

+++ Calculate Employee Bonus@@67877d61dc107486ad5c274f/code.c
@@ -14,7 +14,7 @@ int main(){
     }
     for(int i=0;i<num;i++){
         if(employee[i].salary<50000){
-            employee[i].bonus=0.1*(float)employee[i].salary;
+            employee[i].bonus=0.1*employee[i].salary;
         }else{
             employee[i].bonus=(float)(5*employee[i].salary)/100;
         }
